+++ src/core/event_loop.rs
@@ -75,11 +75,14 @@ fn main_task<B: Backend + std::io::Write>(
         }
     };
 
-    // Check if there is a current song playing on startup
-    if let Some((_, song)) = ctx.find_current_song_in_queue()
-        && let Some(command) = &ctx.config.on_song_change {
-            let mut env = create_env(&ctx, std::iter::empty());
-            run_external(command.clone(), env);
+    // Execute on_song_change at startup if
+    // configured and current song is available.
+    if ctx.config.exec_on_song_change_at_start
+        && let Some((_, song)) = ctx.find_current_song_in_queue()
+        && let Some(command) = &ctx.config.on_song_change
+    {
+        let mut env = create_env(&ctx, std::iter::empty());
+        run_external(command.clone(), env);
     }
 
     match ctx.status.state {